Epoll: remove connections from epoll when they are closed
When Cleanup() is called, the selectable is closed, but it still exists
in the epoll. In addition, events may still be pending for it (due to
races). This leads to the "Epoll broken" situation.

Although this situation is resolved be reconstructing all the epoll
events, that's a bit of overhead. Avoid this by explicitly removing the
event from the epoll, by registering a callback on the Selectable
object.

This also fixes a potential race condition, in case the file descriptor
is reused for another socket, the epoll may still trigger for an old
pending event on the old file descriptor.