Short-circuit CreateTunnels
CreateTunnels is only relevant if backends are configured. When running
as frontend only, it ends up logging "Not sure which servers to contact,
making no changes." all the time, which is spurious and confusing.

Rather than just eliding that message, skip the CreateTunnels function
entirely when no backends are configured.