Tunnel.HandlePageKiteResponse: avoid spurious logging of empty server
HandlePageKiteResponse will log the server capabilities. However, this
function is also called for a front-end, which doesn't have a server
configured. This leads to lots of irrelevant messages in the log
"FE=x.x.x.x:x; ports=; protocols=; raw_ports=".

Probably, for a frontend, HandlePageKiteResponse is irrelevant. However,
in the future, new functionality may be added to it, so we don't want to
skip the function for a front-end entirely.

Instead, skip the log message if server name is not set.

The default server name is 'x.x.x.x:x', which is a bit big to compare
against. Replace it with something simpler. Since the server name is
expected to be a string containing a colon in some other places,
initialize it to ':' instead. This would lead to messages like "FE=:;
..." in case something gets logged for it after all. That indicates just
as clearly that the server is not set.